+++ packages/pwa-kit-runtime/src/ssr/server/test_fixtures/extensions/another-extension/setup-server.js
@@ -4,18 +4,9 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+import {ApplicationExtension} from '../../../extensibility'
 
-class AnotherExtension {
-    options
-
-    constructor(options) {
-        this.options = options
-    }
-
-    getName() {
-        return 'AnotherExtension'
-    }
-
+class AnotherExtension extends ApplicationExtension {
     extendApp(app) {
         app.get('/another-extension', (req, res) => {
             res.send('test')