+++ packages/pwa-kit-runtime/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v4.0.0-dev (Jun 21, 2024)
+- Replace `IApplicationExtension` with `ApplicationExtension` abstract class. [#2019](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2019)
 
 ## v3.7.0 (Aug 07, 2024)